Add automated API docs generation and deployment workflow (#38)
This commit introduces a new Makefile target, `apidocs`, which organizes
the generation of API documentation using Dokka. Additionally, a GitHub
Actions workflow is added to automatically build and publish the
documentation to GitHub Pages upon a release or manual trigger. These
changes streamline the documentation process, ensuring up-to-date docs
are easily accessible.

Co-authored-by: Konstantin Pavlov <{ID}+{username}@users.noreply.github.com>